+++ mani_skill/agents/__init__.py
@@ -1,1 +1,3 @@
-from .registration import REGISTERED_AGENTS
+from .base_agent import BaseAgent
+from .multi_agent import MultiAgent
+from .registration import REGISTERED_AGENTS, register_agent
