+++ mani_skill/utils/sapien_utils.py
@@ -3,6 +3,7 @@ Utilities that work with the simulation / SAPIEN
 """
 from __future__ import annotations
 
+import sys
 from typing import TYPE_CHECKING, Dict, List, Tuple, TypeVar
 
 import numpy as np
@@ -10,15 +11,16 @@ import sapien
 import sapien.physx as physx
 import sapien.render
 import sapien.wrapper.urdf_loader
+import torch
+from sapien.utils import Viewer
 
+from mani_skill.render import SAPIEN_RENDER_SYSTEM
 from mani_skill.utils.geometry.rotation_conversions import matrix_to_quaternion
 from mani_skill.utils.structs.pose import Pose
 
 if TYPE_CHECKING:
     from mani_skill.utils.structs.actor import Actor
-    from mani_skill.envs.scene import ManiSkillScene
-
-import torch
+    from mani_skill.sensors.camera import CameraConfig
 
 from mani_skill.utils.structs.types import Array, Device
 
@@ -115,7 +117,6 @@ def parse_urdf_config(config_dict: dict) -> Dict:
 
     Args:
         config_dict (dict): a dict containing link physical properties.
-        scene (ManiSkillScene): the simulation scene
 
     Returns:
         Dict: urdf config passed to `sapien.URDFLoader.load`.
@@ -436,3 +437,42 @@ def is_state_dict_consistent(state_dict: dict):
                     if v.shape[0] != batch_size:
                         return False
     return True
+
+
+def create_viewer(viewer_camera_config: CameraConfig):
+    """Creates a viewer with the given camera config"""
+    if SAPIEN_RENDER_SYSTEM == "3.0":
+        sapien.render.set_viewer_shader_dir(
+            viewer_camera_config.shader_config.shader_pack
+        )
+        if viewer_camera_config.shader_config.shader_pack[:2] == "rt":
+            sapien.render.set_ray_tracing_denoiser(
+                viewer_camera_config.shader_config.shader_pack_config[
+                    "ray_tracing_denoiser"
+                ]
+            )
+            sapien.render.set_ray_tracing_path_depth(
+                viewer_camera_config.shader_config.shader_pack_config[
+                    "ray_tracing_path_depth"
+                ]
+            )
+            sapien.render.set_ray_tracing_samples_per_pixel(
+                viewer_camera_config.shader_config.shader_pack_config[
+                    "ray_tracing_samples_per_pixel"
+                ]
+            )
+        viewer = Viewer(
+            resolutions=(viewer_camera_config.width, viewer_camera_config.height)
+        )
+        if sys.platform == "darwin":  # macOS
+            viewer.window.set_content_scale(1)
+    elif SAPIEN_RENDER_SYSTEM == "3.1":
+        # TODO (stao): figure out how shader pack configs can be set at run time
+        viewer = Viewer(
+            resolutions=(viewer_camera_config.width, viewer_camera_config.height),
+            shader_pack=sapien.render.get_shader_pack(
+                viewer_camera_config.shader_config.shader_pack
+            ),
+        )
+
+    return viewer
